move: ProductPriceChangeEvent에 대해서 각 도메인에 대해 공유되는 이벤트이므로 shared/event로 이동
