+++ src/main/java/kitchenpos/products/tobe/domain/Product.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.tobe.domain;
 
 
-import kitchenpos.shared.ProductPriceChangeEvent;
+import kitchenpos.shared.event.ProductPriceChangeEvent;
 import org.springframework.data.domain.AbstractAggregateRoot;
 
 import javax.persistence.*;
