+++ src/main/java/kitchenpos/shared/event/ProductPriceChangeEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.shared;
+package kitchenpos.shared.event;
 
 import java.math.BigDecimal;
 import java.util.UUID;
